requirements.txt
fastapi==0.115.12
uvicorn==0.34.0 
langgraph==0.3.21
langchain-mcp-adapters==0.0.5
langchain_openai==0.3.11
python-multipart==0.0.20
python-dotenv==1.1.0
